requirements.txt
langchain == 0.0.339
python-dotenv
streamlit
tiktoken
faiss-cpu
protobuf~
langchain-experimental
mysql-connector-python
pymysql
google-generativeai
